scene-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.2
daphne==4.1.2
Django==5.1.7
django-cors-headers==4.7.0
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
frozenlist==1.6.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
gunicorn==23.0.0
h11==0.14.0
http_ece==1.2.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
multidict==6.4.3
oauthlib==3.2.2
packaging==24.2
pika==1.3.2
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg==3.2.7
psycopg-binary==3.2.7
psycopg-pool==3.2.6
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dotenv==1.0.1
pywebpush==2.0.3
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
service-identity==24.2.0
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.1
yarl==1.20.0
zope.interface==7.2
